+++ Modules/computeAverageRating.m
@@ -1,8 +1,7 @@
 function R = computeAverageRating(R)
 
-R.Preference.average = mean( ...
-    [R.Preference.round(1).rating;
-     R.Preference.round(2).rating], ...
-    1);
+ratings = vertcat(R.Preference.round.rating);
+
+R.Preference.average = mean(ratings, 1);
 
 end