+++ Modules/runRatingRound.m
@@ -1,25 +1,22 @@
-function R = runRatingRound(R, P, T, roundNumber)
+function R = runRatingRound(R, P, T, roundNumber, imageIndices)
 
-% Generate random presentation order
-order = randperm(P.Images.nImages);
+% Randomize presentation order
+order = imageIndices(randperm(numel(imageIndices)));
 
 % Store presentation order
 R.Preference.round(roundNumber).order = order;
 
-% Preallocate
-R.Preference.round(roundNumber).rating = nan(1, P.Images.nImages);
-R.Preference.round(roundNumber).RT     = nan(1, P.Images.nImages);
-
 % Present images
-for i = 1:P.Images.nImages
+for i = 1:numel(order)
 
     imageIndex = order(i);
 
-    T.progress = i / P.Images.nImages;
+    T.progress = i / numel(order);
 
-    [rating, RT] = runRatingTrial(P.Images.files{imageIndex}, P, T);
+    [rating, RT] = runRatingTrial( ...
+        P.Images.files{imageIndex}, P, T);
 
-    % Store according to image identity (not presentation order)
+    % Store according to image identity
     R.Preference.round(roundNumber).rating(imageIndex) = rating;
     R.Preference.round(roundNumber).RT(imageIndex)     = RT;